requirements.txt
pyserial==3.4
pytz==2019.3